remove header file and clean includes
