+++ Alignment/MillePedeAlignmentAlgorithm/plugins/MillePedeFileConverter.cc
@@ -1,12 +1,46 @@
 // Original Author:  Broen van Besien
 //         Created:  Thu, 19 Mar 2015 18:12:35 GMT
 
-#include "Alignment/MillePedeAlignmentAlgorithm/plugins/MillePedeFileConverter.h"
+/*
+ * This producer will take a given (binary) file and embed it in the output
+ * edm/root file.
+ *
+ * The file is added as a fileblob to a vector of fileblobs that exists on the
+ * level of the run.
+ * The operation is performed during the endRunProduce phase.
+ *
+ * The file can later be extracted again by the corresponding module
+ * MillePedeFileExtractor.
+ *
+ * For more info about the possible parameters, type:
+ * edmPluginHelp -p MillePedeFileConverter
+ */
+
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/global/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "CondFormats/Common/interface/FileBlobCollection.h"
 #include "FWCore/Framework/interface/LuminosityBlock.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-
 #include <memory>
-#include <fstream>
+
+class MillePedeFileConverter : public edm::global::EDProducer<edm::EndLuminosityBlockProducer> {
+public:
+  explicit MillePedeFileConverter(const edm::ParameterSet&);
+  ~MillePedeFileConverter() override;
+  static void fillDescriptions(edm::ConfigurationDescriptions& descriptions);
+
+private:
+  void produce(edm::StreamID, edm::Event& iEvent, const edm::EventSetup& iSetup) const final;
+  void globalEndLuminosityBlockProduce(edm::LuminosityBlock& iLumi, const edm::EventSetup& iSetup) const final;
+
+  const std::string inputDir_;
+  const std::string inputFileName_;
+  const std::string fileBlobLabel_;
+  const edm::EDPutTokenT<FileBlobCollection> putToken_;
+};
 
 MillePedeFileConverter::MillePedeFileConverter(const edm::ParameterSet& iConfig)
     : inputDir_(iConfig.getParameter<std::string>("fileDir")),
@@ -60,3 +94,6 @@ void MillePedeFileConverter::fillDescriptions(edm::ConfigurationDescriptions& de
       "This is the generic cfi file for the "
       "MillePedeFileConverter");
 }
+
+// define this as a plug-in
+DEFINE_FWK_MODULE(MillePedeFileConverter);
